requirements.txt
Django==1.5.1
requests
django-json-field
django-extensions

